Fix Setup middleware
