+++ src/Http/Support/CmsToolConfigurer.php
@@ -5,7 +5,8 @@ namespace Takemo101\CmsTool\Http\Support;
 use Slim\App as Slim;
 use Slim\Middleware\MethodOverrideMiddleware;
 use Takemo101\Chubby\Http\Configurer\DefaultSlimConfigurer;
-use Takemo101\CmsTool\Http\Middleware\Setup;
+use Takemo101\CmsTool\Http\Middleware\CacheControl;
+use Takemo101\CmsTool\Http\Middleware\SetupProcess;
 
 class CmsToolConfigurer extends DefaultSlimConfigurer
 {
@@ -19,7 +20,8 @@ class CmsToolConfigurer extends DefaultSlimConfigurer
     {
         $slim = parent::configure($slim);
 
-        $slim->add(Setup::class);
+        $slim->add(SetupProcess::class);
+        $slim->add(CacheControl::class);
         $slim->add(new MethodOverrideMiddleware());
 
         return $slim;
